scope: Refactor all Description: initialized a new project with the same name
Expected behavior:
should able to reduce boiler plate code
