--- README.md
@@ -1,4 +1,0 @@
-# fordastore-api
-fordastore core api used in handling requests and sql queries
-
-> :bulb: This readme should contain all the interfaces definition.
