--- src/main/java/com/splitscale/api/fordastore/db/DBPropsLoader.java
@@ -1,10 +1,0 @@
-package com.splitscale.api.fordastore.db;
-
-public interface DBPropsLoader {
-
-  String getUrl();
-
-  String getUsername();
-
-  String getPassword();
-}
