--- src/main/java/com/splitscale/api/fordastore/db/SqlConnectable.java
@@ -1,8 +1,0 @@
-package com.splitscale.api.fordastore.db;
-
-import java.sql.Connection;
-import java.sql.SQLException;
-
-public interface SqlConnectable {
-  public Connection getConnection() throws SQLException;
-}
