--- src/main/java/com/splitscale/api/fordastore/user/UserController.java
@@ -1,13 +1,0 @@
-package com.splitscale.api.fordastore.user;
-
-public interface UserController {
-
-  public UserInfo getUser(String JSONgetPayload);
-
-  public void addUser(UserInfoImpl userInfo);
-
-  public void deleteUser(UserInfoImpl userInfo);
-
-  public void updateUser(UserInfoImpl userInfo);
-
-}
