--- src/main/java/com/splitscale/api/fordastore/user/UserInfo.java
@@ -1,11 +1,0 @@
-package com.splitscale.api.fordastore.user;
-
-public interface UserInfo {
-  public String getUsername();
-
-  public String getPassword();
-
-  public void setUsername(String username);
-
-  public void setPassword(String password);
-}
