--- src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1,17 +1,0 @@
-{"properties": [
-  {
-    "name": "mysql.password",
-    "type": "java.lang.String",
-    "description": "password for mysql database"
-  },
-  {
-    "name": "mysql.url",
-    "type": "java.lang.String",
-    "description": "A description for 'mysql.url'"
-  },
-  {
-    "name": "mysql.username",
-    "type": "java.lang.String",
-    "description": "A description for 'mysql.username'"
-  }
-]}
