+++ src/main/resources/application.properties
@@ -1,8 +1,1 @@
-spring.datasource.driver-class-name=com.mysql.jdbc.Driver
-spring.datasource.url=jdbc:mysql://localhost:3307/fordaStore?allowPublicKeyRetrieval=true&useSSL=false
-spring.datasource.username=root
-spring.datasource.password=splitscale
 
-mysql.url=jdbc:mysql://localhost:3307/fordaStore?allowPublicKeyRetrieval=true&useSSL=false
-mysql.username=root
-mysql.password=splitscale
