--- src/test/java/com/splitscale/api/fordastore/FordastoreApplicationTests.java
@@ -1,18 +1,0 @@
-package com.splitscale.api.fordastore;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-class FordastoreApplicationTests {
-
-  @Test
-  @DisplayName("fordastore package")
-  void someTest() {
-    boolean item1 = true;
-    boolean item2 = true;
-
-    assertEquals(item1, item2);
-  }
-}
